fix(tests): seed tax config once in fixture init to prevent parallel race
All PlaceOrderTests ran SeedTaxConfigAsync per-test against the same shared
brand slugs, racing each other under TUnit's parallel execution. Move the
seeding into IntegrationTestBase.InitializeAsync — it runs once before any
tests start — and remove the redundant per-test calls.

Also reverts the exception-based workaround in TaxConfigurationService.
