+++ src/backend/TheMillionthFoodOrderApp.Application/TaxConfiguration/TaxConfigurationService.cs
@@ -27,21 +27,12 @@ public sealed class TaxConfigurationService(ITaxConfigurationRepository reposito
         {
             var config = Domain.TaxConfiguration.TaxConfiguration.Create();
             config.UpdateRates(rates);
-            try
-            {
-                await repository.AddAsync(config, cancellationToken);
-                await repository.SaveChangesAsync(cancellationToken);
-                return MapToResponse(config);
-            }
-            catch (InvalidOperationException)
-            {
-                // Race condition: a concurrent request created the config between our
-                // GetAsync check and AddAsync. Fall through to update the existing record.
-                existing = await repository.GetAsync(cancellationToken);
-            }
+            await repository.AddAsync(config, cancellationToken);
+            await repository.SaveChangesAsync(cancellationToken);
+            return MapToResponse(config);
         }
 
-        var updated = await repository.ReplaceRatesAsync(existing!.Id, c => c.UpdateRates(rates), cancellationToken);
+        var updated = await repository.ReplaceRatesAsync(existing.Id, c => c.UpdateRates(rates), cancellationToken);
         return MapToResponse(updated);
     }
